chore: create stream_ext module
