+++ tokio-stream/src/stream_ext/all.rs
