+++ tokio-stream/src/stream_ext/any.rs
