+++ tokio-stream/src/stream_ext/collect.rs
