+++ tokio-stream/src/stream_ext/fold.rs
