+++ tokio-stream/src/stream_ext/fuse.rs
