+++ tokio-stream/src/stream_ext/map.rs
