+++ tokio-stream/src/stream_ext/next.rs
