+++ tokio-stream/src/stream_ext/take_while.rs
